Add MapActionEndpointConventionBuilder
