+++ src/Http/Routing/src/Builder/MapActionEndpointConventionBuilder.cs
@@ -1,0 +1,33 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Collections.Generic;
+
+namespace Microsoft.AspNetCore.Builder
+{
+    /// <summary>
+    /// Builds conventions that will be used for customization of MapAction <see cref="EndpointBuilder"/> instances.
+    /// </summary>
+    public sealed class MapActionEndpointConventionBuilder : IEndpointConventionBuilder
+    {
+        private readonly List<IEndpointConventionBuilder> _endpointConventionBuilders;
+
+        internal MapActionEndpointConventionBuilder(List<IEndpointConventionBuilder> endpointConventionBuilders)
+        {
+            _endpointConventionBuilders = endpointConventionBuilders;
+        }
+
+        /// <summary>
+        /// Adds the specified convention to the builder. Conventions are used to customize <see cref="EndpointBuilder"/> instances.
+        /// </summary>
+        /// <param name="convention">The convention to add to the builder.</param>
+        public void Add(Action<EndpointBuilder> convention)
+        {
+            foreach (var endpointConventionBuilder in _endpointConventionBuilders)
+            {
+                endpointConventionBuilder.Add(convention);
+            }
+        }
+    }
+}
