+++ src/Http/Routing/src/Builder/MapActionEndpointRouteBuilderExtensions.cs
@@ -20,8 +20,8 @@ namespace Microsoft.AspNetCore.Builder
         /// </summary>
         /// <param name="endpoints">The <see cref="IEndpointRouteBuilder"/> to add the route to.</param>
         /// <param name="action">The delegate executed when the endpoint is matched.</param>
-        /// <returns>A <see cref="IEndpointConventionBuilder"/> that can be used to further customize the endpoint.</returns>
-        public static IEndpointConventionBuilder MapAction(
+        /// <returns>An <see cref="IEndpointConventionBuilder"/> that can be used to further customize the endpoint.</returns>
+        public static MapActionEndpointConventionBuilder MapAction(
             this IEndpointRouteBuilder endpoints,
             Delegate action)
         {
@@ -61,7 +61,6 @@ namespace Microsoft.AspNetCore.Builder
                         endpointBuilder.Metadata.Add(attribute);
                     }
 
-
                     endpointBuilder.DisplayName = routeName ?? pattern;
 
                     ((RouteEndpointBuilder)endpointBuilder).Order = routeOrder ?? defaultOrder;
@@ -75,25 +74,7 @@ namespace Microsoft.AspNetCore.Builder
                 throw new InvalidOperationException("Action must have a pattern. Is it missing a Route attribute?");
             }
 
-            return new CompositeEndpointConventionBuilder(conventionBuilders);
-        }
-
-        private class CompositeEndpointConventionBuilder : IEndpointConventionBuilder
-        {
-            private readonly List<IEndpointConventionBuilder> _endpointConventionBuilders;
-
-            public CompositeEndpointConventionBuilder(List<IEndpointConventionBuilder> endpointConventionBuilders)
-            {
-                _endpointConventionBuilders = endpointConventionBuilders;
-            }
-
-            public void Add(Action<EndpointBuilder> convention)
-            {
-                foreach (var endpointConventionBuilder in _endpointConventionBuilders)
-                {
-                    endpointConventionBuilder.Add(convention);
-                }
-            }
+            return new MapActionEndpointConventionBuilder(conventionBuilders);
         }
     }
 }
